Flere opdateringer. Men jeg har ramt et problem med oprettelse af konto - Google har ændret deres OpenID betydeligt, det vil kræve undersøgelse at klare det.

+++ WingtipToys/Models/OrderDetail.cs
@@ -7,6 +7,6 @@ namespace WingtipToys.Models {
         public string Username { get; set; }
         public int ProductId { get ; set; }
         public int Quantity { get; set; }
-        public double? UnitPrice { get ; set; }
+        public decimal? UnitPrice { get ; set; }
         }
     }
